requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.4
django-bootstrap4==23.2
django-extensions==3.2.3
django-filter==23.2
Pillow==10.0.0
python-decouple==3.8
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
